treewide: move everything from shared/nixosModules to modules, add roles options
The idea of this whole refactor is to be able to simplify hosts
configurations, that are getting a little too messy and repeated for my
sake. The main idea is, have everything defined as modules (except
host-specific configuration such as nvidia drivers for my desktop for
example). We are going to have everything go under modules, and the goal
of this refactor is to get rid of the config directory, which makes for
an awkward mix between import-based and module-based.