shared/modules/hjem: update environment module to use `~/.profile` instead
I decided to go with the simpler solution of using `~/.profile`, as that
is the only option that will ensure that the environment will be set
correctly. If you launch a more barebones compositor such as Hyprland
from your tty, since it won't be launched by systemd it will not inherit
the environment.