packages/mcuxpressoide: fix build
Fixed the package's build process, as I was not building all of the
pieces correctly, and it seems to mostly work now (also removed the need
to use the Eclipse wrapping function). What I'm trying to figure out now
is why it won't find LinkServer when first starting the app, as it seems
to look for it in the wrong directory? Changing this manually in
preferences seems to fix it though, but it takes ages to start redlink
and it does hang the entire app in the process. That said, it seems to
be mostly built correctly now!