treewide: separate desktop option between name and type
This will be used when other compositors will be added to the
configuration, as otherwise it will get rather messy. It also allows to
only enable programs/services such as the ones under `gnome.nix`, which
would not be needed when using a DE as those will be already set for the
user.