treewide: use resolveAndFilter
This is the first step towards using the synaptic standard: we are
implementing a `resolveAndFilter` function, which is meant to replace
Nix's `default.nix` functionality.

The idea behind this is that every file under the modules given to
`resolveAndFilter` will be imported as modules, which will simplify the
general config and allow for a completely arbitrary file-tree. The
downside to this approach is, we cannot use files as Nix expressions
anymore (eg. `gtk/style.nix`, which simply returned a string of the css
interpolated with Nix variables).

This is not a big deal, as we can simply rewrite it to turn it into a
module! We could also add special handling of Nix expressions if we
would want to keep that, by filtering out files starting with `_`, but
that's unnecessary for now.

We also move our shared modules directly into `parts`, as this requires
special handling as they will be exported modules (we just need to
return their path, and that wasn't possible as we currently import
EVERYTHING from `modules`).