modules/monitors: update positions API; add assertions
Considering we want to be able to use different compositors, the
positions API had to be reworked: the main idea before was to work with
Hyprland, and pass the position string directly, but considering I want
to be able to use Niri in the future, that would not work as their
syntax is different.

Now, position is nullable, in case we want to use auto monitor
positioning (this will then be mapped in the compositor's
configuration). If it is set, both x and y will have to be set as the
configuration would otherwise be invalid (I'd rather have it explicitely
set rather than default to 0).